feat: full ±999,999,999-year date range — TCK 3880/3880 (100%)
The last 2 failing TCK scenarios (Temporal10 [9]/[10]) need dates at
year ±999,999,999. Two hard limits stood in the way: PropertyValue::Date
stored i32 epoch days (caps ~±5.8M years), and all calendar math went
through chrono::NaiveDate (caps ±262k years).

- PropertyValue::Date(i32) -> Date(i64). Wire-compatible with existing
  files: postcard varints don't encode integer width, and the property-
  index key encoding was already 8-byte sign-flipped i64 — no format
  bump, no re-encode.
- temporal.rs calendar core rewritten as hand-rolled proleptic-Gregorian
  integer math (Hinnant's civil-calendar algorithms): ymd/epoch-day
  conversion, ISO week dates, ordinal/quarter forms, month arithmetic
  with end-of-month clamping, truncation. chrono remains only for now()
  capture and named-IANA-zone resolution (inherently bounded by the IANA
  database itself; fixed offsets are pure arithmetic).
- duration.between core moved to i128 nanosecond totals — the full
  range's span (~6.3e25 ns) overflows i64 even though its second count
  fits.
- Parser accepts ISO 8601 expanded years ([+-]YYYYYYYYY-MM-DD, week and
  ordinal forms included); formatting round-trips them with an explicit
  sign outside 0000..=9999. localdatetime() now also accepts a date-only
  string (midnight), per the TCK.
- Year range validated at every constructor (±999,999,999, java.time's
  range, matching real Cypher).

Tests: civil core cross-checked against chrono over ±2.7 millennia of
epoch days (components, weekdays, ISO weeks); expanded-year roundtrip,
range/validity boundaries (1900/2000 leap rules); both TCK scenarios as
unit + end-to-end tests including storage roundtrip. Full TCK:
3880/3880, all suites green, python/capi/cli bindings compile clean.

+++ marsdb-graph/src/model.rs
@@ -49,8 +49,12 @@ pub enum PropertyValue {
     /// and keeps comparison a plain integer compare. Conversion to/from
     /// calendar year/month/day and ISO-8601 text lives in `marsdb-query`
     /// (`temporal.rs`), not here -- this crate only stores the value, it
-    /// doesn't know Cypher's date grammar/semantics.
-    Date(i32),
+    /// doesn't know Cypher's date grammar/semantics. `i64`, not `i32`:
+    /// Cypher's full year range (±999_999_999, ISO 8601 expanded years)
+    /// reaches ±365 billion epoch days, past `i32`. Wire-compatible with
+    /// values written as `i32`: postcard varints don't encode the width,
+    /// and the index key encoding was already 8-byte (see `index.rs`).
+    Date(i64),
     /// An ISO-8601 duration (Cypher's `DURATION` type), kept in Neo4j's
     /// own four-component normalized form rather than as a single scalar
     /// -- months and days are *not* fungible with each other or with
